+++ chartlets.js/src/lib/component/Registry.ts
@@ -8,11 +8,21 @@ export interface Registry {
   /**
    * Register a React component that renders a Chartlets component.
    *
+   * `component` must be a functional React component with at
+   * least the following two component props:
+   *
+   *   - `type: string`: your component's type name.
+   *      This will be the same as the `type` used for registration.
+   *   - `onChange: ComponentChangeHandler`: an event handler
+   *     that your component may call to signal change events.
+   *
+   * Both props are always be present plus. The component may also
+   * be passed any other props provided by the contribution.
+   *
+   * @param type The Chartlets component's unique type name.
    * @param component A functional React component.
-   * @param type The Chartlets component's type name.
-   *   If not provided, `component.name` is used.
    */
-  register(component: FC<ComponentProps>, type?: string): () => void;
+  register(type: string, component: FC<ComponentProps>): () => void;
 
   /**
    * Lookup the component of the provided type.
@@ -31,8 +41,7 @@ export interface Registry {
 export class RegistryImpl implements Registry {
   private components = new Map<string, FC<ComponentProps>>();
 
-  register(component: FC<ComponentProps>, type?: string): () => void {
-    type = type || component.name;
+  register(type: string, component: FC<ComponentProps>): () => void {
     const oldComponent = this.components.get(type);
     this.components.set(type, component);
     return () => {
@@ -56,7 +65,7 @@ export class RegistryImpl implements Registry {
 /**
  * The Chartly component registry.
  *
- * Use `registry.register(C)` to register your own component `C`.
+ * Use `registry.register("C", C)` to register your own component `C`.
  *
  * `C` must be a functional React component with at least the following
  * two properties: